Fix: 컬러/이미지 토글시 ImageOption 컴포넌트의 업로드한 이미지 데이터가 초기화되는 문제 해결
